Restructure chat item component
