+++ src/components/chats/ChatItem/ChatItem.tsx
@@ -1,19 +1,18 @@
-import { getPostQuery } from '@/services/api/query'
+import AddressAvatar from '@/components/AddressAvatar'
+import Button from '@/components/Button'
+import LinkText from '@/components/LinkText'
 import { useSendEvent } from '@/stores/analytics'
 import { truncateAddress } from '@/utils/account'
 import { cx } from '@/utils/class-names'
 import { getTimeRelativeToNow } from '@/utils/date'
-import { getExplorerUrl } from '@/utils/explorer'
-import { getIpfsContentUrl } from '@/utils/ipfs'
 import { generateRandomColor } from '@/utils/random-colors'
-import { truncateText } from '@/utils/text'
 import Linkify from 'linkify-react'
 import { ComponentProps, useReducer } from 'react'
 import { IoCheckmarkDoneOutline, IoCheckmarkOutline } from 'react-icons/io5'
-import AddressAvatar from '../AddressAvatar'
-import Button from '../Button'
-import LinkText from '../LinkText'
-import Modal, { ModalFunctionalityProps } from '../Modal'
+import CheckMarkExplanationModal, {
+  CheckMarkModalVariant,
+} from './CheckMarkExplanationModal'
+import RepliedMessagePreview from './RepliedMessagePreview'
 
 export type ChatItemProps = Omit<ComponentProps<'div'>, 'children'> & {
   text: string
@@ -151,124 +150,5 @@ export default function ChatItem({
         cid={cid}
       />
     </div>
-  )
-}
-
-type RepliedMessagePreviewProps = ComponentProps<'div'> & {
-  replyTo: string
-  originalMessage: string
-}
-const MINIMUM_REPLY_CHAR = 20
-const MAXIMUM_REPLY_CHAT = 100
-function RepliedMessagePreview({
-  replyTo,
-  originalMessage,
-  ...props
-}: RepliedMessagePreviewProps) {
-  const { data } = getPostQuery.useQuery(replyTo)
-  if (!data) {
-    // TODO: show loading
-    return null
-  }
-
-  const replySender = data.struct.ownerId
-  const replySenderColor = generateRandomColor(replySender)
-
-  let showedText = data.content?.body ?? ''
-  if (originalMessage.length < MINIMUM_REPLY_CHAR) {
-    showedText = truncateText(showedText, MINIMUM_REPLY_CHAR)
-  }
-
-  return (
-    <div
-      {...props}
-      className={cx(
-        'flex flex-col overflow-hidden border-l-2 pl-2 text-sm',
-        props.className
-      )}
-      style={{ borderColor: replySenderColor, ...props.style }}
-    >
-      <span style={{ color: replySenderColor }}>
-        {truncateAddress(data?.struct.ownerId ?? '')}
-      </span>
-      <span className='overflow-hidden overflow-ellipsis whitespace-nowrap'>
-        {showedText}
-      </span>
-    </div>
-  )
-}
-
-type CheckMarkDescData = {
-  blockNumber?: number
-  cid?: string
-}
-type CheckMarkExplanationModalProps = ModalFunctionalityProps & {
-  variant: CheckMarkModalVariant
-} & CheckMarkDescData
-
-const variants = {
-  recording: {
-    title: 'Recording message',
-    icon: <IoCheckmarkOutline className='text-text-muted' />,
-    desc: () => (
-      <span>
-        Your message is being processed. In the next few seconds, it will be
-        saved to the blockchain, after which you will see this icon:{' '}
-        <IoCheckmarkDoneOutline className='inline text-text' />
-      </span>
-    ),
-  },
-  recorded: {
-    title: 'Message recorded',
-    icon: <IoCheckmarkDoneOutline />,
-    desc: ({ blockNumber, cid }) => (
-      <span>
-        Your censorship-resistant message has been stored on{' '}
-        <LinkText
-          href={cid ? getIpfsContentUrl(cid) : 'https://ipfs.tech/'}
-          openInNewTab
-          variant='primary'
-        >
-          IPFS
-        </LinkText>{' '}
-        and the{' '}
-        <LinkText
-          href={
-            blockNumber
-              ? getExplorerUrl(blockNumber)
-              : 'https://subsocial.network/'
-          }
-          openInNewTab
-          variant='primary'
-        >
-          Subsocial blockchain
-        </LinkText>
-        .
-      </span>
-    ),
-  },
-} satisfies {
-  [key: string]: {
-    title: string
-    icon: JSX.Element
-    desc: (data: CheckMarkDescData) => JSX.Element
-  }
-}
-type CheckMarkModalVariant = keyof typeof variants
-function CheckMarkExplanationModal({
-  variant,
-  blockNumber,
-  cid,
-  ...props
-}: CheckMarkExplanationModalProps) {
-  const { title, icon, desc } = variants[variant]
-
-  return (
-    <Modal {...props} title={title} withCloseButton>
-      <div className='flex flex-col items-center'>
-        <div className='my-6 flex justify-center text-8xl'>{icon}</div>
-        <p className='text-text-muted'>{desc({ blockNumber, cid })}</p>
-      </div>
-    </Modal>
   )
 }
