+++ src/components/chats/ChatItem/index.tsx
@@ -1,0 +1,4 @@
+import ChatItem from './ChatItem'
+
+export * from './ChatItem'
+export default ChatItem
